LPD-52709 Convert HashedFilesRegistry to a POJO
