+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/js/importmaps/extender/FrontendJsWebDynamicJSImportMapsContributor.java
@@ -15,7 +15,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.io.Writer;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -59,7 +62,18 @@ public class FrontendJsWebDynamicJSImportMapsContributor
 		throws IOException {
 	}
 
-	@Reference
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_hashedFilesRegistry = new HashedFilesRegistry(bundleContext);
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_hashedFilesRegistry.close();
+
+		_hashedFilesRegistry = null;
+	}
+
 	private HashedFilesRegistry _hashedFilesRegistry;
 
 	@Reference
